transplant jQuery code to TableOfContents component
